requirements.txt
streamlit
feedparser
requests
Pillow